Kernel: See below.
Fix BUG inside ATA driver: Due to bad cast on lba select.
Add ToolBox APIs for kernel GUI.

Userland: See below.

Worked a bit on System.Core, just wrapped Thread.hxx into the System namespace.

+++ Private/Builtins/ATA/Defines.hxx
@@ -136,10 +136,10 @@ HCore::Void drv_ata_select(HCore::UInt16 Bus);
 
 HCore::Boolean drv_ata_wait_io(HCore::UInt16 IO);
 
-HCore::Void drv_ata_read(HCore::UInt32 Lba, HCore::UInt16 IO, HCore::UInt8 Master, HCore::Char* Buf,
+HCore::Void drv_ata_read(HCore::UInt64 Lba, HCore::UInt16 IO, HCore::UInt8 Master, HCore::Char* Buf,
                    HCore::SizeT SectorSz, HCore::SizeT Size);
 
-HCore::Void drv_ata_write(HCore::UInt32 Lba, HCore::UInt16 IO, HCore::UInt8 Master, HCore::Char* Buf,
+HCore::Void drv_ata_write(HCore::UInt64 Lba, HCore::UInt16 IO, HCore::UInt8 Master, HCore::Char* Buf,
                     HCore::SizeT SectorSz, HCore::SizeT Size);
 
 #endif // ifdef __KERNEL__
