+++ Private/NewBoot/Source/HEL/AMD64/BootATA.cxx
@@ -131,7 +131,7 @@ ATAInit_Retry:
   return true;
 }
 
-Void boot_ata_read(UInt32 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
+Void boot_ata_read(UInt64 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
                    SizeT SectorSz, SizeT Size) {
   UInt8 Command = (!Master ? 0xE0 : 0xF0);
 
@@ -140,9 +140,9 @@ Void boot_ata_read(UInt32 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
   Out8(IO + ATA_REG_HDDEVSEL, (Command) | (((Lba) >> 24) & 0xF));
   Out8(IO + ATA_REG_SEC_COUNT0, SectorSz);
 
-  Out8(IO + ATA_REG_LBA0, (UInt8)(Lba));
-  Out8(IO + ATA_REG_LBA1, (UInt8)(Lba) >> 8);
-  Out8(IO + ATA_REG_LBA2, (UInt8)(Lba) >> 16);
+  Out8(IO + ATA_REG_LBA0, (Lba));
+  Out8(IO + ATA_REG_LBA1, (Lba) >> 8);
+  Out8(IO + ATA_REG_LBA2, (Lba) >> 16);
 
   Out8(IO + ATA_REG_COMMAND, ATA_CMD_READ_PIO);
 
@@ -164,7 +164,7 @@ Void boot_ata_read(UInt32 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
   }
 }
 
-Void boot_ata_write(UInt32 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
+Void boot_ata_write(UInt64 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
                     SizeT SectorSz, SizeT Size) {
   UInt8 Command = (!Master ? 0xE0 : 0xF0);
 
@@ -173,9 +173,9 @@ Void boot_ata_write(UInt32 Lba, UInt16 IO, UInt8 Master, CharacterTypeUTF8* Buf,
   Out8(IO + ATA_REG_HDDEVSEL, (Command) | (((Lba) >> 24) & 0xF));
   Out8(IO + ATA_REG_SEC_COUNT0, SectorSz);
 
-  Out8(IO + ATA_REG_LBA0, (UInt8)(Lba));
-  Out8(IO + ATA_REG_LBA1, (UInt8)(Lba) >> 8);
-  Out8(IO + ATA_REG_LBA2, (UInt8)(Lba) >> 16);
+  Out8(IO + ATA_REG_LBA0, (Lba));
+  Out8(IO + ATA_REG_LBA1, (Lba) >> 8);
+  Out8(IO + ATA_REG_LBA2, (Lba) >> 16);
 
   Out8(IO + ATA_REG_COMMAND, ATA_CMD_WRITE_PIO);
